requirements.txt
# requirements.txt

flask
Flask-Cors
Flask-Login
Flask-PyMongo
pymongo
python-dotenv